Tests/LibWeb/Text/expected/css/deferred-font-face-does-not-block-load.txt

Window load fired after stylesheet load: true
DeferredFontFace found: true
DeferredFontFace status: unloaded
